chore(python): move dev keys from BootloaderV2Image [no changelog]
